requirements.txt
Django~=1.9                 # WEB framework
apns-clerk==0.2.*           # APNS client
python-gcm==0.4.*           # Google Cloud Messaging client
kombu==3.0.*                # Messaging library
